[refactor] Composition 폴더 정리
